[FIX] stock: inventory adjustement one product
Usecase to reproduce:
- Create a product tracked by serial number
- Do an inventory adjustement for it with the selected option 'one
product only'
- Add a new line with a serial number
- Save

Error due to missing product_id field at the inventory.line creation.
It's due to readonly set on the view. It will select by default the
inventory product but it will not send it to the server.

Add force_save in order to send the field even if it's readonly

Fix #39902

closes odoo/odoo#39935

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>